+++ Wall.h
@@ -9,7 +9,7 @@ class Wall :
 	public StaticThing
 {
 public:
-	Wall(sf::RenderWindow &win, int x1, int y1,int x2, int y2);
+	Wall(sf::RenderWindow &win, double x1, double y1, double x2, double y2);
 	~Wall();
 	void display(void);
 	void addDoor(Door* newDoor);
